requirementsO.txt
fastapi
uvicorn
requests
boto3
httpx
mysql-connector-python
